rag-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12   # for file uploads (UploadFile)
PyJWT==2.8.0              # for JWT token handling
python-jose[cryptography]==3.3.0  # for JWT with cryptography
requests>=2.32.5          # for fetching Google public keys
redis==5.0.1              # for chat history storage
